# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Jannat Events public website and content management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export type EventStatus = typeof EventStatus[keyof typeof EventStatus];


export const EventStatus = {
  UPCOMING: 'UPCOMING',
  PAST: 'PAST',
  DRAFT: 'DRAFT',
} as const;

export type MediaType = typeof MediaType[keyof typeof MediaType];


export const MediaType = {
  IMAGE: 'IMAGE',
  VIDEO: 'VIDEO',
} as const;

export interface Media {
  id: number;
  type: MediaType;
  url: string;
  /** @nullable */
  thumbnail?: string | null;
  altText: string;
  /** @nullable */
  albumId?: number | null;
  /** @nullable */
  eventId?: number | null;
  sortOrder: number;
  createdAt: string;
}

export interface MediaInput {
  type: MediaType;
  /** @minLength 1 */
  url: string;
  /** @nullable */
  thumbnail?: string | null;
  altText: string;
  /** @nullable */
  albumId?: number | null;
  /** @nullable */
  eventId?: number | null;
  sortOrder?: number;
}

export interface MediaUpdate {
  altText?: string;
  /** @nullable */
  thumbnail?: string | null;
  /** @nullable */
  albumId?: number | null;
  /** @nullable */
  eventId?: number | null;
  sortOrder?: number;
}

export interface Event {
  id: number;
  title: string;
  slug: string;
  date: string;
  startTime: string;
  /** @nullable */
  endTime?: string | null;
  venue: string;
  address: string;
  city: string;
  province: string;
  description: string;
  /** @nullable */
  poster?: string | null;
  /** @nullable */
  ticketUrl?: string | null;
  /** @nullable */
  instagramUrl?: string | null;
  status: EventStatus;
  featured: boolean;
  /** @nullable */
  ageRestriction?: string | null;
  /** @nullable */
  dressCode?: string | null;
  /** @nullable */
  organizer?: string | null;
  /** @nullable */
  seoTitle?: string | null;
  /** @nullable */
  seoDescription?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface EventInput {
  /** @minLength 1 */
  title: string;
  /** @nullable */
  slug?: string | null;
  date: string;
  startTime: string;
  /** @nullable */
  endTime?: string | null;
  venue: string;
  address: string;
  city: string;
  province: string;
  description: string;
  /** @nullable */
  poster?: string | null;
  /** @nullable */
  ticketUrl?: string | null;
  /** @nullable */
  instagramUrl?: string | null;
  status?: EventStatus;
  featured?: boolean;
  /** @nullable */
  ageRestriction?: string | null;
  /** @nullable */
  dressCode?: string | null;
  /** @nullable */
  organizer?: string | null;
  /** @nullable */
  seoTitle?: string | null;
  /** @nullable */
  seoDescription?: string | null;
}

export interface EventUpdate {
  /** @minLength 1 */
  title?: string;
  slug?: string;
  date?: string;
  startTime?: string;
  /** @nullable */
  endTime?: string | null;
  venue?: string;
  address?: string;
  city?: string;
  province?: string;
  description?: string;
  /** @nullable */
  poster?: string | null;
  /** @nullable */
  ticketUrl?: string | null;
  /** @nullable */
  instagramUrl?: string | null;
  status?: EventStatus;
  featured?: boolean;
  /** @nullable */
  ageRestriction?: string | null;
  /** @nullable */
  dressCode?: string | null;
  /** @nullable */
  organizer?: string | null;
  /** @nullable */
  seoTitle?: string | null;
  /** @nullable */
  seoDescription?: string | null;
}

export interface Album {
  id: number;
  title: string;
  slug: string;
  description: string;
  /** @nullable */
  coverImage?: string | null;
  /** @nullable */
  eventId?: number | null;
  /** @nullable */
  eventTitle?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  date?: string | null;
  published: boolean;
  createdAt: string;
}

export type EventDetail = Event & ({
  gallery?: Media[];
  album?: Album | null;
});

export interface AlbumInput {
  /** @minLength 1 */
  title: string;
  /** @nullable */
  slug?: string | null;
  description: string;
  /** @nullable */
  coverImage?: string | null;
  /** @nullable */
  eventId?: number | null;
  published?: boolean;
}

export interface AlbumUpdate {
  /** @minLength 1 */
  title?: string;
  slug?: string;
  description?: string;
  /** @nullable */
  coverImage?: string | null;
  /** @nullable */
  eventId?: number | null;
  published?: boolean;
}

export type AlbumDetail = Album & {
  media?: Media[];
};

export type ContactInputInquiryType = typeof ContactInputInquiryType[keyof typeof ContactInputInquiryType];


export const ContactInputInquiryType = {
  General: 'General',
  Event: 'Event',
  Sponsorship: 'Sponsorship',
  Partnership: 'Partnership',
  Artist_Booking: 'Artist Booking',
  Media: 'Media',
  Other: 'Other',
} as const;

export interface ContactInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @minLength 1 */
  subject: string;
  inquiryType: ContactInputInquiryType;
  /** @minLength 1 */
  message: string;
}

export type ContactMessageStatus = typeof ContactMessageStatus[keyof typeof ContactMessageStatus];


export const ContactMessageStatus = {
  NEW: 'NEW',
  READ: 'READ',
  ARCHIVED: 'ARCHIVED',
} as const;

export type ContactMessage = ContactInput & {
  id: number;
  status: ContactMessageStatus;
  createdAt: string;
};

export interface HomePage {
  /** @nullable */
  heroImage?: string | null;
  /** @nullable */
  heroVideo?: string | null;
  heroHeadline: string;
  heroSubtitle: string;
  ctaText: string;
  ctaUrl: string;
  /** @nullable */
  featuredEventId?: number | null;
  featuredEvent?: Event | null;
  gallery: Media[];
}

export interface HomePageUpdate {
  /** @nullable */
  heroImage?: string | null;
  /** @nullable */
  heroVideo?: string | null;
  heroHeadline?: string;
  heroSubtitle?: string;
  ctaText?: string;
  ctaUrl?: string;
  /** @nullable */
  featuredEventId?: number | null;
}

export interface SiteSettings {
  instagramUrl: string;
  /** @nullable */
  facebookUrl?: string | null;
  /** @nullable */
  tiktokUrl?: string | null;
  /** @nullable */
  contactEmail?: string | null;
  cities: string[];
}

export interface SiteSettingsUpdate {
  instagramUrl?: string;
  /** @nullable */
  facebookUrl?: string | null;
  /** @nullable */
  tiktokUrl?: string | null;
  /** @nullable */
  contactEmail?: string | null;
  cities?: string[];
}

export interface AdminOverview {
  upcomingEvents: number;
  pastEvents: number;
  publishedAlbums: number;
  totalMedia: number;
  unreadMessages: number;
  recentEvents: Event[];
}

export interface UploadRequest {
  /** @minLength 1 */
  name: string;
  /**
     * @minimum 1
     * @maximum 524288000
     */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadResponse {
  uploadURL: string;
  objectPath: string;
}

/**
 * Invalid input
 */
export type BadRequestResponse = Error;

/**
 * Resource not found
 */
export type NotFoundResponse = Error;

export type CityParameter = string;

export type SearchParameter = string;

export type EventStatusParameter = EventStatus;

export type MediaTypeParameter = MediaType;

export type ListPublicEventsParams = {
city?: CityParameter;
status?: EventStatusParameter;
};

export type ListAdminEventsParams = {
status?: EventStatusParameter;
search?: SearchParameter;
};

export type ListAdminAlbumsParams = {
search?: SearchParameter;
};

export type ListAdminMediaParams = {
type?: MediaTypeParameter;
search?: SearchParameter;
};