# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Jannat Events public website and content management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /public/events:
    get:
      operationId: listPublicEvents
      tags: [public]
      summary: List published events
      parameters:
        - $ref: "#/components/parameters/City"
        - $ref: "#/components/parameters/EventStatus"
      responses:
        "200":
          description: Published event list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
  /public/events/{slug}:
    get:
      operationId: getPublicEvent
      tags: [public]
      summary: Get a published event by slug
      parameters:
        - $ref: "#/components/parameters/Slug"
      responses:
        "200":
          description: Event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventDetail"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /public/albums/{slug}:
    get:
      operationId: getPublicAlbum
      tags: [public]
      summary: Get a published album by slug
      parameters:
        - $ref: "#/components/parameters/Slug"
      responses:
        "200":
          description: Album
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlbumDetail"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /contact:
    post:
      operationId: submitContact
      tags: [public]
      summary: Submit a contact inquiry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Contact message received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactMessage"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request a presigned upload URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadRequest"
      responses:
        "200":
          description: Presigned upload data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResponse"
components:
  parameters:
...
    Slug:
      name: slug
      in: path
      required: true
      schema:
        type: string
    City:
      name: city
      in: query
      required: false
      schema:
        type: string
...
    EventStatus:
      name: status
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/EventStatus"
...
  responses:
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
    EventStatus:
      type: string
      enum: [UPCOMING, PAST, DRAFT]
    MediaType:
      type: string
      enum: [IMAGE, VIDEO]
    Media:
      type: object
      properties:
        id: { type: integer }
        type: { $ref: "#/components/schemas/MediaType" }
        url: { type: string }
        thumbnail: { type: ["string", "null"] }
        altText: { type: string }
        albumId: { type: ["integer", "null"] }
        eventId: { type: ["integer", "null"] }
        sortOrder: { type: integer }
        createdAt: { type: string, format: date-time }
      required: [id, type, url, altText, sortOrder, createdAt]
...
    Event:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        slug: { type: string }
        date: { type: string, format: date }
        startTime: { type: string }
        endTime: { type: ["string", "null"] }
        venue: { type: string }
        address: { type: string }
        city: { type: string }
        province: { type: string }
        description: { type: string }
        poster: { type: ["string", "null"] }
        ticketUrl: { type: ["string", "null"], format: uri }
        instagramUrl: { type: ["string", "null"], format: uri }
        status: { $ref: "#/components/schemas/EventStatus" }
        featured: { type: boolean }
        ageRestriction: { type: ["string", "null"] }
        dressCode: { type: ["string", "null"] }
        organizer: { type: ["string", "null"] }
        seoTitle: { type: ["string", "null"] }
        seoDescription: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, title, slug, date, startTime, venue, address, city, province, description, status, featured, createdAt, updatedAt]
...
    EventDetail:
      allOf:
        - $ref: "#/components/schemas/Event"
        - type: object
          properties:
            gallery: { type: array, items: { $ref: "#/components/schemas/Media" } }
            album: { anyOf: [{ $ref: "#/components/schemas/Album" }, { type: "null" }] }
    Album:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        slug: { type: string }
        description: { type: string }
        coverImage: { type: ["string", "null"] }
        eventId: { type: ["integer", "null"] }
        eventTitle: { type: ["string", "null"] }
        city: { type: ["string", "null"] }
        date: { type: ["string", "null"], format: date }
        published: { type: boolean }
        createdAt: { type: string, format: date-time }
      required: [id, title, slug, description, published, createdAt]
...
    AlbumDetail:
      allOf:
        - $ref: "#/components/schemas/Album"
        - type: object
          properties:
            media: { type: array, items: { $ref: "#/components/schemas/Media" } }
    ContactInput:
      type: object
      required: [name, email, subject, inquiryType, message]
      properties:
        name: { type: string, minLength: 1 }
        email: { type: string, format: email }
        phone: { type: ["string", "null"] }
        subject: { type: string, minLength: 1 }
        inquiryType: { type: string, enum: [General, Event, Sponsorship, Partnership, "Artist Booking", Media, Other] }
        message: { type: string, minLength: 1 }
    ContactMessage:
      allOf:
        - $ref: "#/components/schemas/ContactInput"
        - type: object
          properties:
            id: { type: integer }
            status: { type: string, enum: [NEW, READ, ARCHIVED] }
            createdAt: { type: string, format: date-time }
          required: [id, status, createdAt]
...
    UploadRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name: { type: string, minLength: 1 }
        size: { type: integer, minimum: 1, maximum: 524288000 }
        contentType: { type: string, minLength: 1 }
    UploadResponse:
      type: object
      properties:
        uploadURL: { type: string, format: uri }
        objectPath: { type: string }
      required: [uploadURL, objectPath]