# lib/api-zod/src/generated/types/contactInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Jannat Events public website and content management API
 * OpenAPI spec version: 0.1.0
 */
import type { ContactInputInquiryType } from './contactInputInquiryType';

export interface ContactInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @minLength 1 */
  subject: string;
  inquiryType: ContactInputInquiryType;
  /** @minLength 1 */
  message: string;
}