# lib/api-zod/src/generated/types/media.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Jannat Events public website and content management API
 * OpenAPI spec version: 0.1.0
 */
import type { MediaType } from './mediaType';

export interface Media {
  id: number;
  type: MediaType;
  url: string;
  /** @nullable */
  thumbnail?: string | null;
  altText: string;
  /** @nullable */
  albumId?: number | null;
  /** @nullable */
  eventId?: number | null;
  sortOrder: number;
  createdAt: Date;
}